allen_academy_login/requirements.txt
Django==4.2.6
djangorestframework==3.15.1
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
